leetcode/0142.linked-list-cycle-ii/src/linked-list-cycle-ii_test.go
package src

import (
	"testing"
)

func Test_detectCycle(t *testing.T) {
	type args struct {
		head *ListNode
	}

	node := &ListNode{
		Val: 0,
		Next: &ListNode{
			Val:  0,
			Next: nil,
		},
	}
	node.Next.Next = node

	tests := []struct {
		name string
		args args
		want *ListNode
	}{
		{
			name: "detectCycle--1",
			args: args{
				head: &ListNode{
					Val: 0,
					Next: &ListNode{
						Val: 1,
						Next: &ListNode{
							Val: 2,
							Next: &ListNode{
								Val:  0,
								Next: node,
							},
						},
					},
				},
			},
			want: node,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := detectCycle(tt.args.head); got != tt.want {
				t.Errorf("detectCycle() = %v, want %v", got, tt.want)
			}
		})
	}
}
